Moved headers to include directory for inclusion in derived projects
